+++ src/libs/videoPuzzle/AudioManager.js
@@ -2,8 +2,8 @@ export class AudioManager {
   constructor() {
     this.musicOn = true
     this.sfxOn = true
-    PIXI.loader.resources['mus_TimeToDream'].sound.loop = true
-    PIXI.sound.volume('mus_TimeToDream', 0.3)
+    PIXI.loader.resources['mus_infinityOcean'].sound.loop = true
+    PIXI.sound.volume('mus_infinityOcean', 0.3)
   }
 
   playSfx(soundRef) {