+++ changelog.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## v.0.7
+## v.1.0-beta
 
 - Changed title screen
-- Change game name to "Fractured Flicks"
+- Changed game name to "Fractured Flicks"
+- Changed background music
 - Pieces now rotate at 180-degree intervals instead of 90 to streamline gameplay
 - Changed production deployment process to use push-dir